(admin) abstract and split node tabs
